Added authorization middleware, and injected userId into all routes

+++ storage/postgre.ts
@@ -3,48 +3,65 @@ import { knex } from '../db/postgre';
 
 const tableName = 'todos';
 
-const listAll = async () => {
-  return knex(tableName)
-    .select();
+const assertUserId = (userId: string) => {
+  if (!userId) {
+    throw new Error('userId must be provided');
+  }
 };
 
-const getById = async (id: string) => {
+const listAll = async (userId: string) => {
+  assertUserId(userId);
+
+  return knex(tableName)
+    .select()
+    .where({ userId });
+};
+
+const getById = async (userId: string, id: string) => {
+  assertUserId(userId);
+
   const list = await knex(tableName)
     .select()
-    .where({ id });
+    .where({ id, userId });
 
   return list[0];
 };
 
-const create = async (item: ItemType) => {
+const create = async (userId: string, item: ItemType) => {
+  assertUserId(userId);
+
   const { id, title, complete } = item;
 
   const list = await knex(tableName)
-    .insert({ id, title, complete })
+    .insert({ id, title, complete, userId })
     .returning('*');
 
   return list[0];
 };
 
-const update = async (item: ItemType) => {
+const update = async (userId: string, item: ItemType) => {
+  assertUserId(userId);
+
   const { id, title, complete } = item;
 
   const list = await knex(tableName)
-    .update({ id, title, complete })
+    .update({ id, title, complete, userId })
     .where({ id })
     .returning('*');
 
   return list[0];
 };
 
-const remove = async (id: string) => {
+const remove = async (userId: string, id: string) => {
+  assertUserId(userId);
+
   if (!id) {
     return;
   }
 
   await knex(tableName)
     .delete()
-    .where({ id });
+    .where({ id, userId });
 };
 
 export {
@@ -52,5 +69,5 @@ export {
   getById,
   create,
   update,
-  remove
-}
+  remove,
+};
